Ver6 0731 pm

+++ application.py
@@ -193,11 +193,11 @@ def login():
     # User reached route via POST (as by submitting a form via POST)
     if request.method == "POST":
         
-        # Forget any userid
-        session.clear()
-
         # Purge shadow table to ensure no errant entries
-        db.execute("DELETE FROM shadow") 
+        db.execute("DELETE FROM shadow WHERE shauserlink = ?", session["user_id"]) 
+
+        # Forget any userid
+        session.clear()        
 
         """
         # NOTE: temporary code to block none "tester" accounts
@@ -275,12 +275,12 @@ def login():
 def logout():
     # NOTE - LOGOUT DEFINITION ORIGINALLY FROM CS50PSET9, WITHOUT MODIFICATION
     """Log user out"""
+    
+    # Purge shadow table to ensure no errant entries
+    db.execute("DELETE FROM shadow WHERE shauserlink = ?", session["user_id"]) 
 
     # Forget any userid
-    session.clear()
-
-    # Purge shadow table to ensure no errant entries
-    db.execute("DELETE FROM shadow") 
+    session.clear()    
 
     # Redirect user to login form
     return redirect("/login")
@@ -296,7 +296,7 @@ def profile():
     session["current_time"] = datetime.now(pytz.utc) #.strftime("%Y/%m/%d %H:%M:%S")
 
     # Purge shadow table to ensure no errant entries
-    db.execute("DELETE FROM shadow") 
+    db.execute("DELETE FROM shadow WHERE shauserlink = ?", session["user_id"]) 
 
     # last_page tracks the last page (either login/index/recallpin/recallall/profile)
     session["last_page"] = "/profile"
@@ -418,7 +418,7 @@ def clearrecords():
     session["current_time"] = datetime.now(pytz.utc) #.strftime("%Y/%m/%d %H:%M:%S")
 
     # Purge shadow table to ensure no errant entries
-    db.execute("DELETE FROM shadow") 
+    db.execute("DELETE FROM shadow WHERE shauserlink = ?", session["user_id"]) 
 
     # last_page tracks the last page (either login/index/recallpin/recallall/profile)
     session["last_page"] = "/profile"
@@ -434,7 +434,7 @@ def deleteentries():
     session["current_time"] = datetime.now(pytz.utc) #.strftime("%Y/%m/%d %H:%M:%S")
 
     # Purge shadow table to ensure no errant entries
-    db.execute("DELETE FROM shadow") 
+    db.execute("DELETE FROM shadow WHERE shauserlink = ?", session["user_id"]) 
 
     # last_page tracks the last page (either login/index/recallpin/recallall/profile)
     session["last_page"] = "/profile"
@@ -492,7 +492,7 @@ def deleteaccount():
     """Deletes all data for user"""
 
     # Purge shadow table to ensure no errant entries
-    db.execute("DELETE FROM shadow") 
+    db.execute("DELETE FROM shadow WHERE shauserlink = ?", session["user_id"]) 
 
     # User reached route via POST (as by submitting a form via POST)
     if request.method == "POST":
@@ -568,20 +568,20 @@ def index():
     session["current_time"] = datetime.now(pytz.utc) #.strftime("%Y/%m/%d %H:%M:%S")
 
     # Purge shadow table to ensure no errant entries
-    db.execute("DELETE FROM shadow") 
+    db.execute("DELETE FROM shadow WHERE shauserlink = ?", session["user_id"]) 
     
     # Create empty list, to populate with *list of dictionaries*
-    allvocabtable = []
+    notpinvocabtable = []
     pinvocabtable = []
     allcount = 25
     pincount = 10
 
     # PRINT TESTS (FOR DEBUGGING)
     #print("test, vocabtable[{}]: ", vocabtable)
-    #print("test, allvocabtable (before): ", allvocabtable)
+    #print("test, notpinvocabtable (before): ", notpinvocabtable)
     #print("test, pinvocabtable (before): ", pinvocabtable)
     
-    # Filter out for allvocabtable
+    # Filter out for notpinvocabtable
     # NOTE: go through all dictionary items within the list that db.execute returns
     for vocabtable_list in vocabtable:
         #print("test, vocabtable_list (list-of-dict): ", vocabtable_list)
@@ -601,15 +601,15 @@ def index():
                         if vt_key == "pin" and vt_value == False:
                             #print("log: pinned=False match")
                             #print("test, vocabtable_list (dict): ", vocabtable_list)
-                            allvocabtable.append(vocabtable_list)
-                            #print("test, allvocabtable.append: ", allvocabtable)
+                            notpinvocabtable.append(vocabtable_list)
+                            #print("test, notpinvocabtable.append: ", notpinvocabtable)
                             # NOTE: subtract 1 from allcount
                             allcount = allcount - 1
             else:
                 # End loop once 25 entries have been detected
                 #print("log: function break due to allcount at ", allcount)
                 break            
-    #print("test, allvocabtable (after): ", allvocabtable)
+    #print("test, notpinvocabtable (after): ", notpinvocabtable)
 
     # Filter out for pinvocabtable
     # NOTE: go through all dictionary items within the list that db.execute returns
@@ -642,7 +642,7 @@ def index():
                 break  
     #print("test, pinvocabtable (after): ", pinvocabtable)
     
-    return render_template("index.html", vocabtable=vocabtable, allvocabtable=allvocabtable, pinvocabtable=pinvocabtable, all_languages=all_languages)
+    return render_template("index.html", vocabtable=vocabtable, notpinvocabtable=notpinvocabtable, pinvocabtable=pinvocabtable, all_languages=all_languages)
 
 @app.route("/recallpin")
 @login_required
@@ -672,7 +672,7 @@ def recallpin():
     session["current_time"] = datetime.now(pytz.utc) #.strftime("%Y/%m/%d %H:%M:%S")
 
     # Purge shadow table to ensure no errant entries
-    db.execute("DELETE FROM shadow") 
+    db.execute("DELETE FROM shadow WHERE shauserlink = ?", session["user_id"]) 
 
     # Create empty list, to populate with *list of dictionaries*
     pinnedvocabtable = []
@@ -728,7 +728,7 @@ def recallall():
     session["current_time"] = datetime.now(pytz.utc) #.strftime("%Y/%m/%d %H:%M:%S")
 
     # Purge shadow table to ensure no errant entries
-    db.execute("DELETE FROM shadow") 
+    db.execute("DELETE FROM shadow WHERE shauserlink = ?", session["user_id"]) 
 
     # Create empty list, to populate with *list of dictionaries*
     fullvocabtable = []
@@ -764,7 +764,7 @@ def input():
     session["current_time"] = datetime.now(pytz.utc) #.strftime("%Y/%m/%d %H:%M:%S")
 
     # Purge shadow table to ensure no errant entries
-    db.execute("DELETE FROM shadow") 
+    db.execute("DELETE FROM shadow WHERE shauserlink = ?", session["user_id"]) 
     
     # User reached route via POST (as by submitting a form via POST)
     if request.method == "POST":
@@ -929,7 +929,7 @@ def review():
         ) 
 
         # Purge shadow table after every successful insertion to vocab table
-        db.execute("DELETE FROM shadow") 
+        db.execute("DELETE FROM shadow WHERE shauserlink = ?", session["user_id"]) 
 
         return redirect("/")
 
@@ -949,7 +949,7 @@ def editentry():
     session["current_time"] = datetime.now(pytz.utc) #.strftime("%Y/%m/%d %H:%M:%S")
 
     # Purge shadow table to ensure no errant entries
-    db.execute("DELETE FROM shadow") 
+    db.execute("DELETE FROM shadow WHERE shauserlink = ?", session["user_id"]) 
 
     # User reached route via POST (as by submitting a form via POST)
     if request.method == "POST":
@@ -1110,6 +1110,22 @@ def revision():
     # Check if shadowcopy exists
     if len(shadowcopy) == 0:
         flash("Internal Server Error. Kindly Repeat Edit.", category="error")
+        # Redirect based on the value of last_page (aside from /login)
+        # last_page tracks the last page (either login/index/recallpin/recallall/profile)
+        #                 
+        # Redirect to index
+        if session["last_page"] == "/":
+            return redirect("/")
+        
+        # Redirect to recallpinned
+        if session["last_page"] == "/recallpin":
+            return redirect("/recallpin")
+        
+        # Redirect to recallall        
+        if session["last_page"] == "/recallall":
+            return redirect("/recallall")        
+        
+        # Redirect to index.html (backup)
         return redirect("/")
     # BUG NOTE: rare issue where user index out of range, unable to reliably reproduce - suspected issue to be purged shadow table due to out-of-order operation:
     """
@@ -1119,7 +1135,7 @@ def revision():
     """
 
     # Purge shadow table to ensure no errant entries
-    db.execute("DELETE FROM shadow") 
+    db.execute("DELETE FROM shadow WHERE shauserlink = ?", session["user_id"]) 
 
     # User reached route via POST (as by submitting a form via POST)
     if request.method == "POST":
@@ -1171,7 +1187,7 @@ def deletecheck():
     session["current_time"] = datetime.now(pytz.utc) #.strftime("%Y/%m/%d %H:%M:%S")
 
     # Purge shadow table to ensure no errant entries
-    db.execute("DELETE FROM shadow") 
+    db.execute("DELETE FROM shadow WHERE shauserlink = ?", session["user_id"]) 
 
     # User reached route via POST (as by submitting a form via POST)
     if request.method == "POST":
@@ -1197,7 +1213,7 @@ def deletion():
     session["current_time"] = datetime.now(pytz.utc) #.strftime("%Y/%m/%d %H:%M:%S")
 
     # Purge shadow table to ensure no errant entries
-    db.execute("DELETE FROM shadow") 
+    db.execute("DELETE FROM shadow WHERE shauserlink = ?", session["user_id"]) 
 
     # User reached route via POST (as by submitting a form via POST)
     if request.method == "POST":
@@ -1270,7 +1286,7 @@ def pinentry():
     session["current_time"] = datetime.now(pytz.utc) #.strftime("%Y/%m/%d %H:%M:%S")
 
     # Purge shadow table to ensure no errant entries
-    db.execute("DELETE FROM shadow") 
+    db.execute("DELETE FROM shadow WHERE shauserlink = ?", session["user_id"]) 
 
     # User reached route via POST (as by submitting a form via POST)
     if request.method == "POST":
@@ -1339,7 +1355,7 @@ def unpinentry():
     session["current_time"] = datetime.now(pytz.utc) #.strftime("%Y/%m/%d %H:%M:%S")
 
     # Purge shadow table to ensure no errant entries
-    db.execute("DELETE FROM shadow") 
+    db.execute("DELETE FROM shadow WHERE shauserlink = ?", session["user_id"]) 
 
     # User reached route via POST (as by submitting a form via POST)
     if request.method == "POST":
